requirements.txt
blinker==1.5
Bootstrap-Flask==2.2.0
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Sitemap==0.4.0
flask-talisman==1.0.0
Flask-WTF==1.1.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
Werkzeug==2.2.3
WTForms==3.0.1
